+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v1.2.0dev
+
+### `Added`
+
+- [#71](https://github.com/nf-core/epitopeprediction/pull/71) - Add support for the non-free netmhc tool family
+
 ## v1.1.0 - Morgenstelle - 2020-10-20
 
 ### `Added`